Add Swing GUI + groups/friends feature
Clients can now add other clients as friends and have private
conversations, receive notifications, create groups, and add users to
those groups so long as they are an admin. Clients can be added to a
group as an admin or as a regular user. All this is implemented with
a new Swing gui with buttons at the top, a group list on the left,
and a friend list on the right. The database on the Server side has
been expanded with new tables to handle this new functionality.

Communication over the network is still handled in the familiar
sender/receiver pattern but now these Threads use object streams
to send a new user defined type that implements the 'Serializable'
interface. This SerializableMessage is extended by many other classes
that are needed by both the Client and the Server while running.
The SerializableMessage has a String type field that all subclasses
use to specify their type. This allows the receiver threads to cast
to the appropriate object upon receiving the SerializableMessage.